Renamed the Invokable class to CallableClass.
