+++ src/Jaxon.php
@@ -281,7 +281,7 @@ class Jaxon
      *
      * @param string        $sClass             The class name
      *
-     * @return \Jaxon\Request\Factory\Invokable\Request
+     * @return \Jaxon\Request\Factory\CallableClass\Request
      */
     public function request($sClassName)
     {
